code/src/EmailToServiceRequest/requirements.txt
Flask
python-docx
pandas
Pillow
PyPDF2
python-magic
python-magic-bin
pytesseract
PyMuPDF